Improve tests and fix some bugs in the design
The design failed to wire the busy indicator to the output bit, this was
revealed when I fixed the test to wait for that bit to come low.

Also, collatz module now leaves internal registers untouched after it
exits the compute phase, this helps to make sense of the waveforms when
debugging.