Rewrite all the computations in a non-racy way
Rewrite the collatz module to be clock-independent. Change all the rest
of the state management to be non-racy as well.

This fixes an off-by-one overshoot of orbit lengths. However, it still
contains a hack that checks for the end of computations by comparing the
iterator to 2 instead of 1. This helps in this specific case when the
second to last value is well known, but would not work in a general
case.