Save space by not holding the full width of path record
Store only the upper 16 bits of the path record. This is enough to
identify interesting inputs that can be recomputed offline.

This also simplifies overflow detection: now it's not really important
if the iterator overflows the 144 bits, it will leave the uppermost bits
tainted, which will be enough to identify the input number as
interesting candidate for further recomputing.